Fully removing tailscale
I found more files related to tailscale. It looks like the apt-get remove did not clean fully.
Also adding the "-f" flag when doing "rm" of a file. Its good to have that flag in case the files are still in use. Else the rm will fail saying that the file is in use.

```
pi@speeder-pad:/$ sudo find . |grep tailscale
./etc/default/tailscaled
./etc/systemd/system/multi-user.target.wants/tailscaled.service
./etc/systemd/system/tailscaled.service
./var/cache/tailscale
./var/cache/apt/archives/tailscale_1.32.2_armhf.deb
./var/lib/systemd/deb-systemd-helper-masked/tailscaled.service
./var/lib/systemd/deb-systemd-helper-enabled/multi-user.target.wants/tailscaled.service
./var/lib/systemd/deb-systemd-helper-enabled/tailscaled.service.dsh-also
./var/lib/dpkg/info/tailscale.postrm
./var/lib/dpkg/info/tailscale.list
```